WebSockets/tests/submissions/Opera/closing-handshake/003.html

<!doctype html>
<title>WebSockets: client sends closing handshake</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../constants.js></script>
<div id=log></div>
<script>
test(function(){
  var ws = new WebSocket(SCHEME_AND_DOMAIN+':8007/echo');
  ws.onmessage = ws.onerror = assertUnreached;
  ws.onopen = function(e) {
    debug(e);
    ws.close();
  }
  ws.onclose = function(e) {
    debug(e);
    assert_equals(e.wasClean, true, 'e.wasClean');
    ws.onclose = assertUnreached;
    setTestTimeout(null);
    setTimeout(end, 50);
  }
});
</script>
